+++ workspace/adi.max326xx/adi.mcu.max326xx/Idle.em.ts
@@ -1,14 +1,24 @@
 import em from '@$$emscript'
 export const $U = em.$declare('MODULE', IdleI)
 
-import * as IdleI from '@em.hal/IdleI.em'
+import * as $R from '@adi.distro.max326xx/REGS.em'
 
-export namespace em$meta {}
+import * as IdleI from '@em.hal/IdleI.em'
+import * as IntrVec from '@em.arch.arm/IntrVec.em'
+
+export namespace em$meta { }
 
 export function em$startup() {
     $['%%b+']
 }
 
-export function exec() {}
+export function exec() {
+    $['%%b:'](1)
+    $['%%b-']
+    IntrVec.PRIMASK_set(1)
+    e$`asm volatile ("wfi")`
+    $['%%b+']
+    IntrVec.PRIMASK_set(0)
+}
 
-export function wakeup() {}
+export function wakeup() { }