+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * `adi.max326xx` distro package with support for `MAX32655FTRHR` board
 * **Setup** and **Board** status-bar items in VS Code
 * **EM&bull;SCRIPT TOUR GUIDE** view in VS Code
+* `emscript genregs` sub-command
 
 ## VERSION-26.1.3
 * optional `inherits` parameter for `em.$declare`